Move interfaces around to split things better
It was weird having the rotator be the source of handles, and the store only
supporting one item. So split them up better. This lets the store be a
repository for multiple keys, and a single rotator process can rotate them.
Apps can also just get handles from the store, without needing to construct a
rotator. This will enable a better privilege separation, where rotation happens
out of band and normal processes don't need the ability to write data.